# Improvements to type marshaling