--- a/src/WinRT.Runtime/Projections/Type.cs
+++ b/src/WinRT.Runtime/Projections/Type.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 using System;
+using System.Collections.Concurrent;
 using WinRT;
 
 namespace ABI.System
@@ -55,7 +56,20 @@ namespace ABI.System
                 }
             }
 
-            return (kind == TypeKind.Custom ? value.AssemblyQualifiedName : TypeNameSupport.GetNameForType(value, TypeNameGenerationFlags.None), kind);
+            return (GetNameForTypeCached(value, kind == TypeKind.Custom), kind);
+        }
+
+        private static readonly ConcurrentDictionary<global::System.Type, string> typeNameCache = new();
+        private static string GetNameForTypeCached(global::System.Type value, bool customKind)
+        {
+            if (customKind)
+            {
+                return typeNameCache.GetOrAdd(value, (type) => type.AssemblyQualifiedName);
+            }
+            else
+            {
+                return typeNameCache.GetOrAdd(value, (type) => TypeNameSupport.GetNameForType(type, TypeNameGenerationFlags.None));
+            }
         }
 
         public static Marshaler CreateMarshaler(global::System.Type value)
@@ -90,7 +104,7 @@ namespace ABI.System
                 return global::System.Type.GetType(name);
             }
 
-            return TypeNameSupport.FindTypeByName(name.AsSpan()).type;
+            return TypeNameSupport.FindTypeByNameCached(name);
         }
 
         public static unsafe void CopyAbi(Marshaler arg, IntPtr dest) =>